+++ subjects/TLCS/90/TLCS-90.reko/TLCS-90_code.c
@@ -666,7 +666,7 @@ void fn22A6(word16 bc, struct Eq_n * ix, word16 wArg05, ptr16 wArg10)
 	ix->bFFFFFFFC = ix->bFFFFFFF7;
 	ix->bFFFFFFFD = 0x00;
 	byte v15_n = wArg10 + (wArg05 + 1);
-	struct Eq_n * hl_n = (uint16) (uint8) v15_n * 0x08 + (uint16) ((uint8) v15_n) + bc;
+	struct Eq_n * hl_n = (uint16) v15_n * 0x08 + (uint16) v15_n + bc;
 	ix->bFFFFFFF8 = hl_n->b0002;
 	ix->bFFFFFFF9 = hl_n->b0003;
 	fn03AA();
@@ -678,7 +678,7 @@ void fn3E2D(word16 bc, struct Eq_n * ix, word16 wArg05, ptr16 wArg10)
 	ix->bFFFFFFFC = ix->bFFFFFFF7;
 	ix->bFFFFFFFD = 0x00;
 	byte v15_n = wArg10 + (wArg05 + 1);
-	struct Eq_n * hl_n = (uint16) (uint8) v15_n * 0x08 + (uint16) ((uint8) v15_n) + bc;
+	struct Eq_n * hl_n = (uint16) v15_n * 0x08 + (uint16) v15_n + bc;
 	ix->bFFFFFFF8 = hl_n->b0002;
 	ix->bFFFFFFF9 = hl_n->b0003;
 	fn03AA();
